fixed package name
